Extract an abstract PooledConnectionProvider, provide default implementation
